pyproject.toml
[project]
name = "pii-redaction-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Clarie Lee",email = "glee2429@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "spacy (>=3.8.5,<4.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "sentence-transformers (>=4.0.2,<5.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "transformers (>=4.50.3,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]

[tool.poetry]
packages = [
    { include = "app" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
